% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api}
\alias{api}
\title{The URL for the Mathable API endpoint being useds}
\format{A character string}
\usage{
netmathtools:::api
}
\description{
The URL for the Mathable API endpoint being useds
}
\keyword{datasets}
