% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_progress.R
\name{calc_progress}
\alias{calc_progress}
\title{Calculate a student's status in their course}
\usage{
calc_progress(course_id, latest_lesson, latest_tryit, end_date)
}
\arguments{
\item{course_id}{The course id for which the student is enrolled, used to
select the correct schedule}

\item{latest_lesson}{The latest lesson that the student has completed}

\item{latest_tryit}{The latest Try It that the student has completed}

\item{end_date}{The end date of the student, must be an R date object}
}
\value{
A data table with columns:
\itemize{
   \item \code{ProgressStatus} The status of the student determined by the
       criteria set by the NetMath office
   \item \code{DaysLeft} The number of days left in the course the student has
   \item \code{DaysBehind} The number of days behind the recommended schedule
       the student is
   \item \code{TryItsBehind} The number of Try Its behind the recommended
       schedule the student is
   \item \code{LessonsBehind} The number of Lessons behind the recommended
       schedule the student is
   \item \code{CurrentPace} The number of Try Its submitted per day the
       student is averaging
   \item \code{CurrentInterp} A plain-english (approximate) interpretation of
       what the student's current pace means about their submission behavior
   \item \code{TryItsLeft} The number of Try Its the student has left in the
       course
   \item \code{NeededPace} The number of Try Its per day the student must
       average in order to complete the course on time
   \item \code{NeededInterp} A plain-english (approximate) interpretation of
       what the \code{NeededPace} means in terms of needed submission behavior
       from the student. Intentionally rounded up to the half day
}
}
\description{
Calculate a student's status in their course
}
\examples{
calc_progress("deployedcourses/uiuc_netmath_math461_r2012_mm",
              latest_lesson = 4,
              latest_tryit = 3,
              end_date = as.Date("12/25/2016", format = "\%m/\%d/\%Y"))
}
