% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mentor_courses.R
\name{get_mentor_courses}
\alias{get_mentor_courses}
\title{Get the courses for which you mentor}
\usage{
get_mentor_courses(h = NULL, user = NULL, passwd = NULL)
}
\arguments{
\item{h}{A valid handle to a Mathable session (logged in and cookies in jar)}

\item{user}{Your username, e.g. <netid>@netmath.illinois.edu}

\item{passwd}{Your password Mathable password}
}
\value{
A data frame
}
\description{
Get the courses for which you mentor
}

