% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Log into Mathable}
\usage{
login(user, passwd)
}
\arguments{
\item{user}{Your username, e.g. <netid>@netmath.illinois.edu}

\item{passwd}{Your password Mathable password}
}
\value{
A list
\itemize{
   \item \code{handle} A curl handle that is our login session, allowing us
       to query the Mathable API
   \item \code{name} The Mentor's name as it appears in Mathable
}
}
\description{
Will log in the user, find the Mentor Name and return a curl handle
}
\examples{
\dontrun{
session <- login("mkemp6@netmath.illinois.edu", "<passwd>")
}
}
