% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedules}
\alias{schedules}
\title{Recommended Schedules for all the NetMath courses}
\format{A data.frame with N rows representing the expected length of the course
\describe{
  \item{Days}{Day number, since the beginning of the course, ranges from 1-N}
  \item{Lesson}{The Lesson expected to be completed on Days = k}
  \item{Try It}{The Try It in Lesson expected to be completed on Days = k}
}}
\usage{
netmathtools:::schedules
}
\description{
A R list structure with named elements of the courses containing schedules with the following layout:
}
\keyword{datasets}

