% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status}
\alias{status}
\title{Progress status table defining the Green - Red criteria}
\format{A data.frame
\describe{
  \item{low}{The low value on the range}
  \item{high}{The high value on the range}
  \item{ProgressStatus}{The designation that should be assigned for that range}
}}
\usage{
netmathtools:::status
}
\description{
A R data.table structure containing the criteria for assiging the progress status
}
\keyword{datasets}
