% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_progress.R
\name{calc_progress}
\alias{calc_progress}
\title{Calculate a student's status in their course}
\usage{
calc_progress(course_id, latest_lesson, latest_tryit, end_date)
}
\arguments{
\item{course_id}{The course id for which the student is enrolled, used to
select the correct schedule}

\item{latest_lesson}{The latest lesson that the student has completed}

\item{latest_tryit}{The latest Try It that the student has completed}

\item{end_date}{The end date of the student}
}
\value{
A data frame
}
\description{
Calculate a student's status in their course
}

