% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_grades.R
\name{get_all_grades}
\alias{get_all_grades}
\title{Get the all grades your students for the courses you mentor}
\usage{
get_all_grades(h = NULL, user = NULL, passwd = NULL, all = FALSE,
  active = TRUE)
}
\arguments{
\item{h}{A valid handle to a Mathable session (logged in and cookies in jar)}

\item{all}{Boolean, should all of the grades be pulled, or just the logged in Mentor's?}

\item{active}{Boolean, should only the active students be pulled?}
}
\value{
A data frame
}
\description{
Get the all grades your students for the courses you mentor
}

