% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grades_csv.R
\name{get_grades_csv}
\alias{get_grades_csv}
\title{Get the grades your students for the courses you mentor}
\usage{
get_grades_csv(course_id, name, h = NULL, user = NULL, passwd = NULL,
  all_students = FALSE, active = FALSE)
}
\arguments{
\item{course_id}{The fully qualified id for the course you'd like to pull grades for}

\item{h}{A valid handle to a Mathable session (logged in and cookies in jar)}

\item{all_students}{Boolean, should all of the grades be pulled, or just the logged in Mentor's?}

\item{active}{Boolean, should only the active students be pulled?}
}
\value{
A data frame
}
\description{
Get the grades your students for the courses you mentor
}

