% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_progress.R
\name{get_student_progress}
\alias{get_student_progress}
\title{Get the grades your students progress summary}
\usage{
get_student_progress(h = NULL, user = NULL, passwd = NULL, all = FALSE,
  active = FALSE, outfile = NULL)
}
\arguments{
\item{h}{A valid handle to a Mathable session (logged in and cookies in jar)}

\item{all}{Boolean, should all of the grades be pulled, or just the logged in Mentor's?}

\item{active}{Boolean, should only the active students be pulled?}

\item{outfile}{A filename to write results to.}
}
\value{
A data frame
}
\description{
Get the grades your students progress summary
}

