% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Log into Mathable}
\usage{
login(user, passwd)
}
\arguments{
\item{user}{Your username, e.g. <netid>@netmath.illinois.edu}

\item{passwd}{Your password Mathable password}
}
\description{
Will log in the user and download cookies to disk
}

